Moved event processing funcitons
