+++ sc/tests/bundle.py
@@ -3,16 +3,21 @@ import os
 import sys
 from email.mime.text import MIMEText
 import time
-from sc.app.functions import *
+from sc.app.eventprocessing import check_new, process_events, process_shakemaps, run_scenario
+from sc.app.grid import create_grid
 from sc.app.inventory import add_facs_to_groups, add_users_to_groups
-from sc.app.productdownload import geo_json
+from sc.app.productdownload import geo_json, download_scenario
+from sc.app.servertestfunctions import system_test
 from sc.app.task import Task
+from sc.app.util import merge_dicts, Clock, SC
 from util import create_fac, create_group, create_user
 from .impact import *
-from .jsonencoders import *
 from .inventory import *
+from .jsonencoders import *
+from .notifications import *
 from .productdownload import *
 from .updates import *
+from .urlopener import *
 
 class SystemTest(unittest.TestCase):
     '''
@@ -175,33 +180,6 @@ class TestDBConnet(unittest.TestCase):
         users = db_returnsSqlAList()
         for user in users:
             self.assertTrue(isinstance(user, Base))
-
-
-class TestMailer(unittest.TestCase):
-    '''
-    Test the connection to the SMTP server
-    '''
-    def test_initMailer(self):
-        '''
-        Fails when the Mailer is unable to initialize. Failure
-        is caused by code errors or proxy access
-        '''
-        m = Mailer()
-    
-    def test_sendMail(self):
-        '''
-        Send an email with Mailer object. Will fail with code errors or
-        failure to reach SMTP server
-        '''
-        m = Mailer()
-        
-        you = self.email
-        msg = MIMEText('This email is a test of your ShakeCast SMTP server')
-        msg['Subject'] = 'ShakeCast SMTP TEST'
-        msg['From'] = m.me
-        msg['To'] = you
-        
-        m.send(msg=msg, you=you)
         
         
 class TestSC(unittest.TestCase):
@@ -257,24 +235,6 @@ class TestTemplateManager(unittest.TestCase):
 
         self.assertTrue(result)
 
-class TestURLOpener(unittest.TestCase):
-    '''
-    Test the URLOpener Object
-    '''
-    def test_initURLOpener(self):
-        '''
-        Test URLOpener initialization
-        '''
-        url_opener = URLOpener()
-        
-    def test_OpenURL(self):
-        '''
-        Open a URL with URLOpener. Will fail with code errors or lack
-        of internet connection
-        '''
-        url_opener = URLOpener()
-        google = url_opener.open('https://www.google.com')
-
 
 class TestClock(unittest.TestCase):
     '''
@@ -297,416 +257,6 @@ class TestClock(unittest.TestCase):
     def test_fromTime(self):
         c = Clock()
         c.from_time(time.time())
-
-class TestGroupGetsNotification(unittest.TestCase):
-
-    @dbconnect
-    def test_forFirstNotification(self, session=None):
-
-        sm1 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 1
-        )
-
-        session.add(sm1)
-
-        group = create_group(name='GLOBAL')
-
-        session.add(group)
-
-        n1 = Notification(
-            shakemap = sm1
-        )
-
-        session.add(n1)
-
-        session.commit()
-
-        # check notification 1
-        has_alert_level, new_inspection, update = check_notification_for_group(
-                                                    group,
-                                                    n1,
-                                                    session=session
-                                                )
-        self.assertTrue(has_alert_level)
-        self.assertTrue(new_inspection)
-        self.assertFalse(update)
-
-        objs = [sm1, group ,n1]
-        for obj in objs:
-            session.delete(obj)
-
-        session.commit()
-
-    @dbconnect
-    def test_updateNotificationNoneNone(self, session=None):
-
-        sm1 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 1
-        )
-
-        sm2 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 2
-        )
-
-        session.add_all([sm1, sm2])
-
-        group = create_group(name='GLOBAL')
-
-        session.add(group)
-
-        n1 = Notification(
-            shakemap = sm1
-        )
-
-        n2 = Notification(
-            shakemap = sm2
-        )
-
-        session.add_all([n1, n2])
-
-        session.commit()
-
-        # check notification 1
-        has_alert_level, new_inspection, update = check_notification_for_group(
-                                                    group,
-                                                    n2,
-                                                    session=session
-                                                )
-        self.assertTrue(has_alert_level)
-        self.assertFalse(new_inspection)
-        self.assertTrue(update)
-
-        objs = [sm1, sm2, n1, n2, group]
-        for obj in objs:
-            session.delete(obj)
-        
-        session.commit()
-
-    @dbconnect
-    def test_changedNotificationNoneGrey(self, session=None):
-
-        sm1 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 1
-        )
-
-        sm2 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 2
-        )
-
-        session.add_all([sm1, sm2])
-
-        group = create_group(name='GLOBAL')
-
-        session.add(group)
-
-        n1 = Notification(
-            shakemap = sm1
-        )
-
-        n2 = Notification(
-            shakemap = sm2
-        )
-
-        session.add_all([n1, n2])
-
-        fs = FacilityShaking(
-            shakemap = sm2,
-            weight = .5
-        )
-
-        session.add(fs)
-
-        session.commit()
-
-        # check notification 1
-        has_alert_level, new_inspection, update = check_notification_for_group(
-                                                    group,
-                                                    n2,
-                                                    session=session
-                                                )
-        self.assertTrue(has_alert_level)
-        self.assertTrue(new_inspection)
-        self.assertTrue(update)
-
-        objs = [sm1, sm2, n1, n2, group]
-        for obj in objs:
-            session.delete(obj)
-        
-        session.commit()
-
-    @dbconnect
-    def test_changedNotificationGreyGreen(self, session=None):
-
-        sm1 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 1
-        )
-
-        sm2 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 2
-        )
-
-        session.add_all([sm1, sm2])
-
-        group = create_group(name='GLOBAL')
-
-        session.add(group)
-
-        n1 = Notification(
-            shakemap = sm1
-        )
-
-        n2 = Notification(
-            shakemap = sm2
-        )
-
-        session.add_all([n1, n2])
-
-        fs1 = FacilityShaking(
-            shakemap = sm1,
-            weight = .1
-        )
-
-        fs2 = FacilityShaking(
-            shakemap = sm2,
-            weight = 1
-        )
-
-        session.add_all([fs1, fs2])
-
-        session.commit()
-
-        # check notification 1
-        has_alert_level, new_inspection, update = check_notification_for_group(
-                                                    group,
-                                                    n2,
-                                                    session=session
-                                                )
-        self.assertTrue(has_alert_level)
-        self.assertTrue(new_inspection)
-        self.assertTrue(update)
-
-        objs = [sm1, sm2, n1, n2, group]
-        for obj in objs:
-            session.delete(obj)
-        
-        session.commit()
-
-    @dbconnect
-    def test_unchangedNotificationGreenGreen(self, session=None):
-        sm1 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 1
-        )
-
-        sm2 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 2
-        )
-
-        session.add_all([sm1, sm2])
-
-        group = create_group(name='GLOBAL')
-
-        session.add(group)
-
-        n1 = Notification(
-            shakemap = sm1
-        )
-
-        n2 = Notification(
-            shakemap = sm2
-        )
-
-        session.add_all([n1, n2])
-
-        fac1 = Facility(
-            facility_id = 1
-        )
-
-        fs1 = FacilityShaking(
-            shakemap = sm1,
-            weight = 1,
-            facility = fac1
-        )
-
-        fs2 = FacilityShaking(
-            shakemap = sm2,
-            weight = 1,
-            facility = fac1
-        )
-
-        session.add_all([fac1, fs1, fs2])
-
-        session.commit()
-
-        # check notification 1
-        has_alert_level, new_inspection, update = check_notification_for_group(
-                                                    group,
-                                                    n2,
-                                                    session=session
-                                                )
-        self.assertTrue(has_alert_level)
-        self.assertFalse(new_inspection)
-        self.assertTrue(update)
-
-        objs = [sm1, sm2, n1, n2, group, fac1]
-        for obj in objs:
-            session.delete(obj)
-        
-        session.commit()
-
-    @dbconnect
-    def test_changedNotificationGreenGreen(self, session=None):
-        sm1 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 1
-        )
-
-        sm2 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 2
-        )
-
-        session.add_all([sm1, sm2])
-
-        group = create_group(name='GLOBAL')
-
-        session.add(group)
-
-        n1 = Notification(
-            shakemap = sm1
-        )
-
-        n2 = Notification(
-            shakemap = sm2
-        )
-
-        session.add_all([n1, n2])
-
-        fac1 = Facility(
-            facility_id = 1
-        )
-
-        fac2 = Facility(
-            facility_id = 2
-        )
-
-        fs1 = FacilityShaking(
-            shakemap = sm1,
-            weight = 1,
-            facility = fac1
-        )
-
-        fs2 = FacilityShaking(
-            shakemap = sm2,
-            weight = 1,
-            facility = fac2
-        )
-
-        session.add_all([fac1, fs1, fs2])
-
-        session.commit()
-
-        # check notification 1
-        has_alert_level, new_inspection, update = check_notification_for_group(
-                                                    group,
-                                                    n2,
-                                                    session=session
-                                                )
-        self.assertTrue(has_alert_level)
-        self.assertTrue(new_inspection)
-        self.assertTrue(update)
-
-        objs = [sm1, sm2, n1, n2, group, fac1, fac2]
-        for obj in objs:
-            session.delete(obj)
-        
-        session.commit()
-
-    @dbconnect
-    def test_changedNotificationLength(self, session=None):
-        sm1 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 1
-        )
-
-        sm2 = ShakeMap(
-            shakemap_id = 'shakemap1',
-            shakemap_version = 2
-        )
-
-        session.add_all([sm1, sm2])
-
-        group = create_group(name='GLOBAL')
-
-        session.add(group)
-
-        n1 = Notification(
-            shakemap = sm1
-        )
-
-        n2 = Notification(
-            shakemap = sm2
-        )
-
-        session.add_all([n1, n2])
-
-        fac1 = Facility(
-            facility_id = 1
-        )
-
-        fac2 = Facility(
-            facility_id = 2
-        )
-
-        fac3 = Facility(
-            facility_id = 3
-        )
-
-        fs1 = FacilityShaking(
-            shakemap = sm1,
-            weight = 1,
-            facility = fac1
-        )
-
-        fs2 = FacilityShaking(
-            shakemap = sm2,
-            weight = 1,
-            facility = fac2
-        )
-
-        fs3 = FacilityShaking(
-            shakemap = sm2,
-            weight = .1,
-            facility = fac3
-        )
-
-        session.add_all([fac1, fac2, fac3, fs1, fs2, fs3])
-
-        session.commit()
-
-        # check notification 1
-        has_alert_level, new_inspection, update = check_notification_for_group(
-                                                    group,
-                                                    n2,
-                                                    session=session
-                                                )
-        self.assertTrue(has_alert_level)
-        self.assertTrue(new_inspection)
-        self.assertTrue(update)
-
-        objs = [sm1, sm2, n1, n2, group, fac1, fac2, fac3]
-        for obj in objs:
-            session.delete(obj)
-        
-        session.commit()
 
 
 class TestTask(unittest.TestCase):
@@ -974,17 +524,6 @@ class TestFull(unittest.TestCase):
 
         Session.remove()
 
-    def step21_AlchemyEncoder(self):
-        '''
-        Runs through a default use case of the Alchemy Encoder
-        '''
-        session = Session()
-        events = session.query(Event).all()
-        events_json = json.dumps(events, cls=AlchemyEncoder)
-        Session.remove()
-        event_dict = json.loads(events_json)
-        self.assertTrue(len(event_dict) > 0)
-
     def step22_deleteScenario(self):
         session = Session()
         sm = session.query(ShakeMap).first()
@@ -1006,17 +545,6 @@ class TestFull(unittest.TestCase):
             self.assertIsNone(sm)
 
         Session.remove()
-
-    def step23_badScenario(self):
-        result = run_scenario('a_bad_Event_id')
-        self.assertFalse(result['message']['success'])
-
-    def step24_downloadBadScenario(self):
-        result = download_scenario('not_a_real_scenario')
-        self.assertEqual('failed', result['status'])
-
-    def step25_downloadActualScenario(self):
-        download_scenario('bssc2014903_m6p09_se', scenario=True)
 
     @dbconnect
     def step26_groupInspLevel(self, session=None):
