+++ sc/tests/eventprocessing.py
@@ -1,0 +1,8 @@
+import unittest
+
+from sc.app.eventprocessing import *
+
+class TestScenarioDownload(unittest.TestCase):
+    def test_badScenario(self):
+        result = run_scenario('a_bad_Event_id')
+        self.assertFalse(result['message']['success'])
