+++ sc/tests/urlopener.py
@@ -1,0 +1,23 @@
+import unittest
+
+from sc.app.urlopener import URLOpener
+
+class TestURLOpener(unittest.TestCase):
+    '''
+    Test the URLOpener Object
+    '''
+    def test_initURLOpener(self):
+        '''
+        Test URLOpener initialization
+        '''
+        url_opener = URLOpener()
+        
+    def test_OpenURL(self):
+        '''
+        Open a URL with URLOpener. Will fail with code errors or lack
+        of internet connection
+        '''
+        url_opener = URLOpener()
+        google = url_opener.open('https://www.google.com')
+
+        self.assertIsNotNone(google)
